requirements.txt
fastapi
uvicorn
llama-cpp-python
scikit-learn
pandas
numpy
sqlalchemy
matplotlib






